test: reorganize two mixed test projects into Unit/Integration folders
Phase 3 (#3362), first step. Establishes the namespace = folder layout for
the unit/integration split: tests move into top-level Unit/ and Integration/
folders and the namespace gains the matching .Unit / .Integration segment
(e.g. ...ServiceOwner.Api.Tests.Unit.Controllers). Shared helpers
(e.g. PlatformFixture) stay at the project root namespace.

Only mixed-type projects are split; single-type projects keep their layout
since their type is already conveyed by the project and the Category trait.

- Altinn.AccessManagement.ServiceOwner.Api.Tests: 1 unit + 2 integration files.
- Altinn.Authorization.Integration.Tests: 3 unit + 5 integration files.

Pure renames + namespace edits, no test logic changed. Both projects build
and their unit lanes pass (6 and 43 tests).
