+++ src/libs/Altinn.Authorization.Integration/test/Altinn.Authorization.Integration.Tests/Integration/Register/RoleStreamEndpointTest.cs
@@ -1,7 +1,7 @@
 using Altinn.Authorization.Integration.Platform;
 using Altinn.Authorization.Integration.Platform.Register;
 
-namespace Altinn.Authorization.Integration.Tests.Register;
+namespace Altinn.Authorization.Integration.Tests.Integration.Register;
 
 [IntegrationTest]
 public class RoleStreamEndpointTest : IClassFixture<PlatformFixture>
