+++ src/libs/Altinn.Authorization.Integration/test/Altinn.Authorization.Integration.Tests/Unit/RequestComposerTest.cs
@@ -2,7 +2,7 @@
 using System.Text;
 using Altinn.Authorization.Integration.Platform;
 
-namespace Altinn.Authorization.Integration.Tests;
+namespace Altinn.Authorization.Integration.Tests.Unit;
 
 /// <summary>
 /// Pure-unit tests for <see cref="RequestComposer"/> (no external dependencies).
